requirements.txt
asgiref==3.4.1
coverage==6.1.2
Django==3.2.9
django-js-asset==1.2.2
django-mptt==0.13.4
django-nested-inline==0.4.4
django-phonenumber-field==6.0.0
django-timezone-field==4.2.1
djangorestframework==3.12.4
phonenumberslite==8.12.37
python-decouple==3.5
pytz==2021.3
sqlparse==0.4.2
